# docs/swagger/swagger.yaml
definitions:
...
  model.ChangefeedCommonInfo:
    properties:
      checkpoint_time:
        type: string
      checkpoint_tso:
        type: integer
      error:
        $ref: '#/definitions/model.RunningError'
      id:
        type: string
      namespace:
        type: string
      state:
        type: string
      upstream_id:
        type: integer
    type: object
  model.ChangefeedConfig:
    properties:
      changefeed_id:
        type: string
      filter_rules:
        items:
          type: string
        type: array
      force_replicate:
        default: false
        description: if true, force to replicate some ineligible tables
        type: boolean
      ignore_ineligible_table:
        default: false
        type: boolean
      ignore_txn_start_ts:
        items:
          type: integer
        type: array
      mounter_worker_num:
        default: 16
        type: integer
      namespace:
        type: string
      sink_config:
        $ref: '#/definitions/config.SinkConfig'
      sink_uri:
        type: string
      start_ts:
        type: integer
      target_ts:
        type: integer
      timezone:
        default: system
        description: timezone used when checking sink uri
        type: string
    type: object
  model.ChangefeedDetail:
    properties:
      checkpoint_time:
        type: string
      checkpoint_tso:
        type: integer
      create_time:
        type: string
      creator_version:
        type: string
      error:
        $ref: '#/definitions/model.RunningError'
      error_history:
        items:
          type: integer
        type: array
      id:
        type: string
      namespace:
        type: string
      resolved_ts:
        type: integer
      sink_uri:
        type: string
      sort_engine:
        type: string
      start_ts:
        type: integer
      state:
        type: string
      target_ts:
        type: integer
      task_status:
        items:
          $ref: '#/definitions/model.CaptureTaskStatus'
        type: array
      upstream_id:
        type: integer
    type: object
  model.HTTPError:
    properties:
      error_code:
        type: string
      error_msg:
        type: string
    type: object
...
  v2.ChangefeedCommonInfo:
    properties:
      checkpoint_time:
        type: string
      checkpoint_tso:
        type: integer
      error:
        $ref: '#/definitions/model.RunningError'
      id:
        type: string
      namespace:
        type: string
      state:
        type: string
      upstream_id:
        type: integer
    type: object
  v2.ChangefeedConfig:
    properties:
      ca_path:
        type: string
      cert_allowed_cn:
        items:
          type: string
        type: array
      cert_path:
        type: string
      changefeed_id:
        type: string
      key_path:
        type: string
      namespace:
        type: string
      pd_addrs:
        items:
          type: string
        type: array
      replica_config:
        $ref: '#/definitions/v2.ReplicaConfig'
      sink_uri:
        type: string
      start_ts:
        type: integer
      target_ts:
        type: integer
    type: object
...
  v2.EmptyResponse:
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/changefeeds:
    get:
      consumes:
      - application/json
      description: list all changefeeds in cdc cluster
      parameters:
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChangefeedCommonInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: List changefeed
      tags:
      - changefeed
    post:
      consumes:
      - application/json
      description: create a new changefeed
      parameters:
      - description: changefeed config
        in: body
        name: changefeed
        required: true
        schema:
          $ref: '#/definitions/model.ChangefeedConfig'
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Create changefeed
      tags:
      - changefeed
  /api/v1/changefeeds/{changefeed_id}:
    delete:
      consumes:
      - application/json
      description: Remove a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Remove a changefeed
      tags:
      - changefeed
    get:
      consumes:
      - application/json
      description: get detail information of a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChangefeedDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Get changefeed
      tags:
      - changefeed
    put:
      consumes:
      - application/json
      description: Update a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      - description: changefeed config
        in: body
        name: changefeedConfig
        required: true
        schema:
          $ref: '#/definitions/model.ChangefeedConfig'
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Update a changefeed
      tags:
      - changefeed
  /api/v1/changefeeds/{changefeed_id}/pause:
    post:
      consumes:
      - application/json
      description: Pause a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Pause a changefeed
      tags:
      - changefeed
  /api/v1/changefeeds/{changefeed_id}/resume:
    post:
      consumes:
      - application/json
      description: Resume a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Resume a changefeed
      tags:
      - changefeed
...
  /api/v2/changefeeds:
    get:
      consumes:
      - application/json
      description: list all changefeeds in cdc cluster
      parameters:
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.ChangefeedCommonInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: List changefeed
      tags:
      - changefeed
      - v2
    post:
      consumes:
      - application/json
      description: create a new changefeed
      parameters:
      - description: changefeed config
        in: body
        name: changefeed
        required: true
        schema:
          $ref: '#/definitions/v2.ChangefeedConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ChangeFeedInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Create changefeed
      tags:
      - changefeed
      - v2
  /api/v2/changefeeds/{changefeed_id}:
    delete:
      consumes:
      - application/json
      description: Remove a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Remove a changefeed
      tags:
      - changefeed
      - v2
    get:
      consumes:
      - application/json
      description: get detail information of a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ChangeFeedInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Get changefeed
      tags:
      - changefeed
      - v2
    put:
      consumes:
      - application/json
      description: Update a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      - description: changefeed config
        in: body
        name: changefeedConfig
        required: true
        schema:
          $ref: '#/definitions/v2.ChangefeedConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ChangeFeedInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Update a changefeed
      tags:
      - changefeed
      - v2
  /api/v2/changefeeds/{changefeed_id}/pause:
    post:
      consumes:
      - application/json
      description: Pause a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Pause a changefeed
      tags:
      - changefeed
      - v2
  /api/v2/changefeeds/{changefeed_id}/resume:
    post:
      consumes:
      - application/json
      description: Resume a changefeed
      parameters:
      - description: changefeed_id
        in: path
        name: changefeed_id
        required: true
        type: string
      - description: resume config
        in: body
        name: resumeConfig
        required: true
        schema:
          $ref: '#/definitions/v2.ResumeChangefeedConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Resume a changefeed
      tags:
      - changefeed
      - v2
...
swagger: "2.0"